+++ src/card/cards/evolve.rs
@@ -7,3 +7,5 @@ pub const CLOUDFIN_RAPTOR: CardDefinitionId = CardDefinitionId(1537);
 pub const ADAPTIVE_SNAPJAW: CardDefinitionId = CardDefinitionId(1538);
 pub const CROCANURA: CardDefinitionId = CardDefinitionId(1539);
 pub const SHAMBLESHARK: CardDefinitionId = CardDefinitionId(1540);
+pub const ELUSIVE_KRASIS: CardDefinitionId = CardDefinitionId(1541);
+pub const SOULSWORN_SPIRIT: CardDefinitionId = CardDefinitionId(1542);